==> 1_table.Rmd <==
---
title: "Supplementary Material"
author: "Li Kangguo"
date: "2022-08-18"
output: word_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(DescrTab2)
library(tidyverse)

load('../data/omicron_ve.RData')

set.seed(202205)

# data clean --------------------------------------------------------------

# datafile_cont_reg <- datafile_cont_reg |>
#   mutate(vaccine_expire = case_when(vaccine_g == 1 & vaccine_d <= 4*4*7 ~ 'N',
#                                     vaccine_g == 1 & vaccine_d > 4*4*7 ~ 'Y',
#                                     TRUE ~ 'U'))

datafile_cont_infections <- datafile_cont_reg |>
  filter(outcome_s == 1)
table(datafile_cont_infections$vaccine)
table(datafile_cont_infections$vaccine[datafile_cont_infections$vaccine_expire == 'N'])


# participants characteristic -----------------------------------------------

datafile_cont_table <- datafile_cont_reg |>
  mutate(title = paste(outcome_s, vaccine, sep = '_'),
         gender = str_to_title(gender),
         vaccine_produce = factor(vaccine_produce, levels = c('P', 'V', 'PV'))) |>
  select(title, outcome_s, age, age_g, gender, vaccine, vaccine_produce, vaccine_expire)
datafile_cont_table$title[datafile_cont_table$outcome_s == 0] <- '0'
datafile_cont_table$title <- factor(datafile_cont_table$title,
                                    levels = c('1_0', '1_1', '1_2', '1_3', '0'))
datafile_cont_table$gender[datafile_cont_table$gender == 'Missing'] <- NA
group_labels <- list('Unvaccinated cases', '1 dose cases',
                     '2 doses cases', 'Booster dose cases',
                     'Controls')
names(group_labels) <- levels(datafile_cont_table$title)
```

## Table S1
```{r datafile_cont_table, message=FALSE, warning=FALSE}
descr(select(datafile_cont_table, -outcome_s) |> 
        mutate(title = factor(title, labels = c(rep('1', 4), '0'))),
      group = 'title',
      summary_stats_cont = list(
        median = DescrTab2:::.median,
        Q1 = DescrTab2:::.Q1,
        Q3 = DescrTab2:::.Q3
      ),
      reshape_rows = list(
        `Q1 - Q3` = list(args = c("Q1", "Q3"), fun = function(Q1, Q3) { paste0(Q1, "-", Q3) })),
      test_options = list(
        nonparametric = F,
        exact = F
      ),
      format_options = list(
        print_p = F,
        print_CI = F,
        print_Total = F,
        percent_suffix = "",
        combine_median_Q1_Q3 = T,
        Nmiss_row_percent = F,
        omit_missings_in_categorical_var = F,
        omit_missings_in_group = F,
        categorical_missing_percent_mode = "missing_as_regular_category",
        percent_accuracy = 0.1,
        combine_mean_sd=TRUE))
```


## Table S2

```{r, message=FALSE, warning=FALSE}
descr(select(datafile_cont_table, -outcome_s) |> 
        filter(title != 0) |> 
        mutate(title = factor(title, level = c('1_0', '1_1', '1_2', '1_3'))),
      group = 'title',
      group_labels = group_labels,
      summary_stats_cont = list(
        median = DescrTab2:::.median,
        Q1 = DescrTab2:::.Q1,
        Q3 = DescrTab2:::.Q3
      ),
      reshape_rows = list(
        `Q1 - Q3` = list(args = c("Q1", "Q3"), fun = function(Q1, Q3) { paste0(Q1, "-", Q3) })),
      test_options = list(
        nonparametric = F,
        exact = F
      ),
      format_options = list(
        print_p = F,
        print_CI = F,
        print_Total = F,
        percent_suffix = "",
        combine_median_Q1_Q3 = T,
        Nmiss_row_percent = F,
        omit_missings_in_categorical_var = F,
        omit_missings_in_group = F,
        categorical_missing_percent_mode = "missing_as_regular_category",
        percent_accuracy = 0.1,
        combine_mean_sd=TRUE))
```
